python/requirements.txt
more-itertools==8.14.0
mypy==0.971
pycodestyle==2.9.1
pycryptodome==3.15.0
pytest==7.1.3
pytest-cov==3.0.0
pytest-sugar==0.9.5
pytest-xdist==2.5.0
